Switch over to caching ids since successful bundle
It became clear that I couldn't count on a filename coming in
on the error due to the proxying in node-browserify.  Therefore
I'm just keeping a cache of changed files since the last time
bundle completed without a on.error being thrown.  When an error
does occur, the system re-watches those files as they're likely
the culprit.

Addresses issue: #54